honksquad: refactor: split SharedEscalatedGrabSystem into partial files (#856)
Behavior-preserving extraction of the 527-line single file into three
partials by responsibility:

- SharedEscalatedGrabSystem.cs      - Initialize + Update lifecycle.
- SharedEscalatedGrabSystem.Stages.cs   - TryEscalate / SetStage / GetStage /
  HasStage / ClearEscalation / TryResist / DropStage state machine.
- SharedEscalatedGrabSystem.Handlers.cs - the event subscriptions.

Duplication and SRP cleanups:
- Merge OnPullerDamaged and OnVanillaPullerDamaged into one
  OnDamageCheckDrop on PullerComponent that branches on GrabStateComponent.
- Extract the recurring TryComp<PullableComponent> + TryStopPull pattern
  into StopPullIfPuller.
- Extract the shared portal/buckle/death "clear-on-conflict" shape into
  ClearGrabAndStopPull.
- Table-drive the stage->locale-key popup map via
  GrabStateComponent.StagePopupKeys instead of an inline switch.

Covered by EscalatedGrabTest.cs; no behavior change.
